[DAGE-64] Refactor of rre-embeddings into rre-tools/embedding-model-evaluator Some changes to pyproject.toml files in all packages: - centralized common imports - centralized test runs (to be done from `rre-tools/` folder) - mypy has one common file in `rre-tools/` folder, exactly like ruff
Executed:
- ruff (that has been centralized)
- mypy (that has been centralized)
- unit test (now the ignore of DeprecationWarning happens only in `embedding-model-evaluator/tests/` folder)
