--- rre-embeddings/mypy.ini
@@ -1,46 +1,0 @@
-[mypy]
-python_version = 3.10
-warn_return_any = True
-warn_unused_configs = True
-disallow_untyped_defs = True
-disallow_incomplete_defs = True
-check_untyped_defs = True
-disallow_untyped_decorators = True
-no_implicit_optional = True
-warn_redundant_casts = True
-warn_unused_ignores = True
-warn_no_return = True
-warn_unreachable = True
-
-# Third-party library specific settings
-[mypy-langchain.*]
-ignore_missing_imports = True
-
-[mypy-langchain_core.*]
-ignore_missing_imports = True
-
-[mypy-langchain_openai.*]
-ignore_missing_imports = True
-
-[mypy-langchain_google_genai.*]
-ignore_missing_imports = True
-
-# Per-module options
-[mypy-main.*]
-check_untyped_defs = True
-
-# Disable strict type checking in tests
-[mypy-tests.*]
-disallow_untyped_defs = False
-
-# Ignore missing imports in __init__.py
-[mypy-__init__.*]
-ignore_errors = True
-
-# Ignore mteb untyped warnings
-[mypy-mteb.*]
-ignore_missing_imports = True
-
-# Ignore datasets untyped warnings
-[mypy-datasets.*]
-ignore_missing_imports = True
