--- rre-embeddings/pyproject.toml
@@ -1,41 +1,0 @@
-[project]
-name = "rre_embeddings"
-version = "0.1.0"
-description = "RRE-Embeddings - A tool for embeddings exact search benchmarking"
-readme = "README.md"
-authors = []
-license = {text = "Apache-2.0"}
-classifiers = [
-    "Development Status :: 3 - Alpha",
-    "Intended Audience :: Science/Research",
-    "License :: OSI Approved :: Apache Software License",
-    "Programming Language :: Python :: 3.10",
-    "Topic :: Scientific/Engineering :: Artificial Intelligence"
-]
-requires-python = ">=3.10"
-dependencies = [
-    "pytest>=8.4.1",
-    "pyyaml>=6.0.2",
-    "python-dotenv>=1.1.1",
-    "mteb>=0.1.0",
-    "pydantic>=2.0",
-    "jsonlines>=4.0.0"
-]
-
-
-[project.scripts]
-rre-embeddings = "src.main:main"
-
-[project.optional-dependencies]
-dev = [
-    "mypy",
-    "types-PyYAML",
-    "ruff"
-]
-
-[tool.setuptools]
-packages = ["src"]
-
-[build-system]
-requires = ["setuptools>=68.0", "wheel"]
-build-backend = "setuptools.build_meta"
