--- rre-embeddings/pytest.ini
@@ -1,5 +1,0 @@
-[pytest]
-filterwarnings =
-    ignore::DeprecationWarning
-
-
