--- rre-embeddings/tests/unit/resources/valid_config.yaml
@@ -1,9 +1,0 @@
-model_id: "sentence-transformers/all-MiniLM-L6-v2"
-task_to_evaluate: "retrieval"
-corpus_path: "tests/unit/resources/data/corpus.jsonl"
-queries_path: "tests/unit/resources/data/queries.jsonl"
-candidates_path: "tests/unit/resources/data/candidates.jsonl"
-relevance_scale: "binary"
-output_dest: "output"
-embeddings_dest: "output/dummy_embeddings"
-
