+++ rre-tools/.gitignore
@@ -17,7 +17,7 @@ uv.lock
 # Local Files and Directories
 .output_dataset
 output/
-resources/
+/resources/
 **/tmp
 
 # Docker
