+++ rre-tools/commons/pyproject.toml
@@ -3,19 +3,9 @@ name = "commons"
 version = "0.1.0"
 description = "Common library for shared functions and classes"
 readme = "README.md"
-requires-python = ">=3.13"
+requires-python = ">=3.10"
 dependencies = [
-    "pydantic>=2.0",
-    "pytest>=8.4.1",
-    "pyyaml>=6.0.2"
 ]
-
-[tool.mypy]
-plugins = ["pydantic.mypy"]
-
-[tool.pydantic-mypy]
-init_typed = true
-warn_required_dynamic_aliases = true
 
 [build-system]
 requires = ["uv_build>=0.8.13,<0.9.0"]
