+++ rre-tools/embedding-model-evaluator/resources/data/candidates.jsonl
