+++ rre-tools/embedding-model-evaluator/src/embedding_model_evaluator/custom_tasks/retrieval_task.py
@@ -4,8 +4,8 @@ from typing import Any
 from mteb.abstasks.AbsTask import TaskMetadata
 from mteb.abstasks.AbsTaskRetrieval import AbsTaskRetrieval
 
-from src.config import Config
-from src.utilities.helper import read_corpus, read_queries, read_candidates
+from embedding_model_evaluator.config import Config
+from embedding_model_evaluator.utilities.helper import read_corpus, read_queries, read_candidates
 
 log = logging.getLogger(__name__)
