+++ rre-tools/embedding-model-evaluator/src/embedding_model_evaluator/utilities/helper.py
