+++ rre-tools/embedding-model-evaluator/tests/conftest.py
@@ -1,0 +1,4 @@
+import warnings
+
+# Since some warnings are raised at import time, I found this solution
+warnings.filterwarnings("ignore", category=DeprecationWarning)
