+++ rre-tools/embedding-model-evaluator/tests/unit/resources/data/candidates.jsonl
