+++ rre-tools/mypy.ini
@@ -1,0 +1,43 @@
+[mypy]
+python_version = 3.13
+exclude = tests/
+plugins = pydantic.mypy
+
+# Strict but reasonable
+disallow_untyped_defs = True
+check_untyped_defs = True
+no_implicit_optional = True
+warn_return_any = True
+show_error_codes = True
+pretty = True
+
+# Allow untyped functions in tests for speed
+[mypy-tests.*]
+disallow_untyped_defs = False
+check_untyped_defs = False
+
+# Ignore missing imports for specific third-party libs
+[mypy-langchain.*]
+ignore_missing_imports = True
+
+[mypy-langchain_core.*]
+ignore_missing_imports = True
+
+[mypy-langchain_openai.*]
+ignore_missing_imports = True
+
+[mypy-langchain_google_genai.*]
+ignore_missing_imports = True
+
+[mypy-mteb.*]
+ignore_missing_imports = True
+
+[mypy-datasets.*]
+ignore_missing_imports = True
+
+[mypy-dotenv.*]
+ignore_missing_imports = True
+
+[pydantic-mypy]
+init_typed = True
+warn_required_dynamic_aliases = True
